apps/api/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
pydantic==2.12.5
httpx==0.28.1
python-dotenv==1.0.1
python-dateutil==2.9.0
langchain==1.2.15
langchain-core==1.2.28
langchain-openai==1.1.12
langgraph==1.1.6
openai==2.31.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
lxml==5.2.1
langchain-google-genai==4.2.1
tavily-python==0.5.1
